refactor: made productCard and productFormFields functions
