+++ package.json
@@ -20,8 +20,8 @@
     "@cypress/webpack-dev-server": "^4.1.0",
     "@icons-pack/react-simple-icons": "^12.8.0",
     "@radix-ui/react-alert-dialog": "^1.1.13",
-    "@radix-ui/react-dialog": "^1.1.13",
     "@radix-ui/react-collapsible": "^1.1.10",
+    "@radix-ui/react-dialog": "^1.1.13",
     "@radix-ui/react-dropdown-menu": "^2.1.14",
     "@radix-ui/react-label": "^2.1.6",
     "@radix-ui/react-popover": "^1.1.13",
@@ -36,7 +36,6 @@
     "@team-golfslag/conflux-api-client": "^2.2.1",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "cypress": "^14.3.3",
     "date-fns": "^4.1.0",
     "dotenv": "^16.5.0",
     "eslint-plugin-headers": "^1.2.1",
@@ -59,6 +58,7 @@
     "@types/react": "^19.1.4",
     "@types/react-dom": "^19.1.5",
     "@vitejs/plugin-react": "^4.4.1",
+    "cypress": "^14.3.3",
     "eslint": "^9.27.0",
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-react-refresh": "^0.4.20",
