+++ src/components/addProductModal.tsx
@@ -13,17 +13,6 @@ import {
   DialogTitle,
   DialogTrigger,
 } from "@/components/ui/dialog";
-import {
-  Select,
-  SelectContent,
-  SelectGroup,
-  SelectItem,
-  SelectLabel,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
 import { Button } from "@/components/ui/button";
 import { Plus } from "lucide-react";
 import {
@@ -35,19 +24,14 @@ import {
 } from "@team-golfslag/conflux-api-client/src/client.ts";
 import { useContext } from "react";
 import { ApiClientContext } from "@/lib/ApiClientContext.ts";
+import ProductFormFields from "@/components/productFormFields.tsx";
+import { ProductFormData } from "@/components/productFormFields.tsx";
 
 type AddWorkModalProps = {
   isOpen: boolean;
   onOpenChange: (open: boolean) => void;
   project: ProjectDTO;
 };
-
-function getEnumKeys<
-  T extends string,
-  TEnumValue extends string | number,
->(enumVariable: { [key in T]: TEnumValue }) {
-  return Object.keys(enumVariable) as Array<T>;
-}
 
 export default function AddProductModal({
   isOpen,
@@ -70,6 +54,13 @@ export default function AddProductModal({
     setUrl("");
     setProductType(undefined);
     setCategory(undefined);
+  };
+
+  const productData: ProductFormData = {
+    url: url,
+    title: productTitle,
+    productType: productType!,
+    productCategory: category!,
   };
 
   const addProduct = async () => {
@@ -120,80 +111,13 @@ export default function AddProductModal({
           <DialogTitle>Add Product</DialogTitle>
           <DialogDescription>Add a product to your project.</DialogDescription>
         </DialogHeader>
-        <div className="grid gap-4 py-4">
-          <div className="grid grid-cols-4 items-center gap-4">
-            <Label htmlFor="title" className="text-right">
-              Title
-            </Label>
-            <Input
-              id="title"
-              value={productTitle}
-              onChange={(e) => setProductTitle(e.target.value)}
-              className="col-span-3"
-            />
-          </div>
-          <div className="grid grid-cols-4 items-center gap-4">
-            <Label htmlFor="url" className="text-right">
-              url
-            </Label>
-            <Input
-              id="url"
-              value={url}
-              onChange={(e) => setUrl(e.target.value)}
-              placeholder="https://doi.org/"
-              className="col-span-3"
-            />
-          </div>
-          <div className="grid grid-cols-4 items-center gap-4">
-            <Label className="col-span-1 text-right">Type</Label>
-            <Select
-              onValueChange={(e) => {
-                setProductType(ProductType[e as keyof typeof ProductType]);
-              }}
-              value={productType}
-            >
-              <SelectTrigger className="col-span-3 w-[180px]">
-                <SelectValue placeholder="Select a product type" />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectGroup>
-                  <SelectLabel>Product types</SelectLabel>
-                  {getEnumKeys(ProductType).map((key, index) => (
-                    <SelectItem key={index} value={ProductType[key]}>
-                      {key}
-                    </SelectItem>
-                  ))}
-                </SelectGroup>
-              </SelectContent>
-            </Select>
-
-            <Label htmlFor="category" className="col-span-1 text-right">
-              Category
-            </Label>
-            <Select
-              onValueChange={(e) => {
-                setCategory(
-                  ProductCategoryType[e as keyof typeof ProductCategoryType],
-                );
-              }}
-              value={category}
-            >
-              <SelectTrigger className="col-span-3 w-[180px]">
-                <SelectValue placeholder="Select a product category type" />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectGroup>
-                  <SelectLabel>Category</SelectLabel>
-                  {getEnumKeys(ProductCategoryType).map((key, index) => (
-                    <SelectItem key={index} value={ProductCategoryType[key]}>
-                      {key}
-                    </SelectItem>
-                  ))}
-                </SelectGroup>
-              </SelectContent>
-            </Select>
-          </div>
-        </div>
+        <ProductFormFields
+          formData={productData}
+          setProductTitle={setProductTitle}
+          setUrl={setUrl}
+          setProductType={setProductType}
+          setCategory={setCategory}
+        />
         <DialogFooter>
           <Button
             variant="outline"
