+++ src/components/editProductModal.tsx
@@ -12,17 +12,6 @@ import {
   DialogHeader,
   DialogTitle,
 } from "@/components/ui/dialog";
-import {
-  Select,
-  SelectContent,
-  SelectGroup,
-  SelectItem,
-  SelectLabel,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
 import { Button } from "@/components/ui/button";
 import {
   ProductCategoryType,
@@ -33,6 +22,9 @@ import {
 } from "@team-golfslag/conflux-api-client/src/client.ts";
 import { useContext, useEffect } from "react";
 import { ApiClientContext } from "@/lib/ApiClientContext.ts";
+import ProductFormFields, {
+  ProductFormData,
+} from "@/components/productFormFields.tsx";
 
 type AddWorkModalProps = {
   isOpen: boolean;
@@ -40,13 +32,6 @@ type AddWorkModalProps = {
   product: ProductDTO;
   project: ProjectDTO;
 };
-
-function getEnumKeys<
-  T extends string,
-  TEnumValue extends string | number,
->(enumVariable: { [key in T]: TEnumValue }) {
-  return Object.keys(enumVariable) as Array<T>;
-}
 
 export default function AddProductModal({
   isOpen,
@@ -73,6 +58,13 @@ export default function AddProductModal({
       setCategory(product.categories[0]);
     }
   }, [product]);
+
+  const productData: ProductFormData = {
+    url: url,
+    title: productTitle,
+    productType: productType!,
+    productCategory: category!,
+  };
 
   const saveEditedProduct = async () => {
     if (!product) return;
@@ -124,80 +116,13 @@ export default function AddProductModal({
             Edit the product of your project.
           </DialogDescription>
         </DialogHeader>
-        <div className="grid gap-4 py-4">
-          <div className="grid grid-cols-4 items-center gap-4">
-            <Label htmlFor="title" className="text-right">
-              Title
-            </Label>
-            <Input
-              id="title"
-              value={productTitle}
-              onChange={(e) => setProductTitle(e.target.value)}
-              className="col-span-3"
-            />
-          </div>
-          <div className="grid grid-cols-4 items-center gap-4">
-            <Label htmlFor="url" className="text-right">
-              url
-            </Label>
-            <Input
-              id="url"
-              value={url}
-              onChange={(e) => setUrl(e.target.value)}
-              placeholder="https://doi.org/"
-              className="col-span-3"
-            />
-          </div>
-          <div className="grid grid-cols-4 items-center gap-4">
-            <Label className="col-span-1 text-right">Type</Label>
-            <Select
-              onValueChange={(e) => {
-                setProductType(ProductType[e as keyof typeof ProductType]);
-              }}
-              value={productType}
-            >
-              <SelectTrigger className="col-span-3 w-[180px]">
-                <SelectValue placeholder="Select a product type" />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectGroup>
-                  <SelectLabel>Product types</SelectLabel>
-                  {getEnumKeys(ProductType).map((key, index) => (
-                    <SelectItem key={index} value={ProductType[key]}>
-                      {key}
-                    </SelectItem>
-                  ))}
-                </SelectGroup>
-              </SelectContent>
-            </Select>
-
-            <Label htmlFor="category" className="col-span-1 text-right">
-              Category
-            </Label>
-            <Select
-              onValueChange={(e) => {
-                setCategory(
-                  ProductCategoryType[e as keyof typeof ProductCategoryType],
-                );
-              }}
-              value={category}
-            >
-              <SelectTrigger className="col-span-3 w-[180px]">
-                <SelectValue placeholder="Select a product category type" />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectGroup>
-                  <SelectLabel>Category</SelectLabel>
-                  {getEnumKeys(ProductCategoryType).map((key, index) => (
-                    <SelectItem key={index} value={ProductCategoryType[key]}>
-                      {key}
-                    </SelectItem>
-                  ))}
-                </SelectGroup>
-              </SelectContent>
-            </Select>
-          </div>
-        </div>
+        <ProductFormFields
+          formData={productData}
+          setProductTitle={setProductTitle}
+          setUrl={setUrl}
+          setProductType={setProductType}
+          setCategory={setCategory}
+        />
         <DialogFooter>
           <Button
             variant="outline"
