+++ src/components/productFormFields.tsx
@@ -1,0 +1,122 @@
+import { Label } from "@/components/ui/label.tsx";
+import { Input } from "@/components/ui/input.tsx";
+import {
+  Select,
+  SelectContent,
+  SelectGroup,
+  SelectItem,
+  SelectLabel,
+  SelectTrigger,
+  SelectValue,
+} from "@/components/ui/select.tsx";
+import {
+  ProductCategoryType,
+  ProductType,
+} from "@team-golfslag/conflux-api-client/src/client.ts";
+
+export interface ProductFormData {
+  title: string;
+  url: string;
+  productType: ProductType;
+  productCategory: ProductCategoryType;
+}
+
+interface ProductFormFieldsProps {
+  formData: ProductFormData;
+  setProductTitle: (title: string) => void;
+  setUrl: (url: string) => void;
+  setProductType: (productType: ProductType) => void;
+  setCategory: (productCategoryType: ProductCategoryType) => void;
+}
+
+function getEnumKeys<
+  T extends string,
+  TEnumValue extends string | number,
+>(enumVariable: { [key in T]: TEnumValue }) {
+  return Object.keys(enumVariable) as Array<T>;
+}
+
+export default function ProductFormFields({
+  formData,
+  setProductTitle,
+  setUrl,
+  setCategory,
+  setProductType,
+}: Readonly<ProductFormFieldsProps>) {
+  return (
+    <div className="grid gap-4 py-4">
+      <div className="grid grid-cols-4 items-center gap-4">
+        <Label htmlFor="title" className="text-right">
+          Title
+        </Label>
+        <Input
+          id="title"
+          value={formData.title}
+          onChange={(e) => setProductTitle(e.target.value)}
+          className="col-span-3"
+        />
+      </div>
+      <div className="grid grid-cols-4 items-center gap-4">
+        <Label htmlFor="url" className="text-right">
+          url
+        </Label>
+        <Input
+          id="url"
+          value={formData.url}
+          onChange={(e) => setUrl(e.target.value)}
+          placeholder="https://doi.org/"
+          className="col-span-3"
+        />
+      </div>
+      <div className="grid grid-cols-4 items-center gap-4">
+        <Label className="col-span-1 text-right">Type</Label>
+        <Select
+          onValueChange={(e) => {
+            setProductType(ProductType[e as keyof typeof ProductType]);
+          }}
+          value={formData.productType}
+        >
+          <SelectTrigger className="col-span-3 w-[180px]">
+            <SelectValue placeholder="Select a product type" />
+          </SelectTrigger>
+          <SelectContent>
+            <SelectGroup>
+              <SelectLabel>Product types</SelectLabel>
+              {getEnumKeys(ProductType).map((key, index) => (
+                <SelectItem key={index} value={ProductType[key]}>
+                  {key}
+                </SelectItem>
+              ))}
+            </SelectGroup>
+          </SelectContent>
+        </Select>
+
+        <Label htmlFor="category" className="col-span-1 text-right">
+          Category
+        </Label>
+        <Select
+          onValueChange={(e) => {
+            setCategory(
+              ProductCategoryType[e as keyof typeof ProductCategoryType],
+            );
+          }}
+          value={formData.productCategory}
+        >
+          <SelectTrigger className="col-span-3 w-[180px]">
+            <SelectValue placeholder="Select a product category type" />
+          </SelectTrigger>
+          <SelectContent>
+            <SelectGroup>
+              <SelectLabel>Category</SelectLabel>
+              {getEnumKeys(ProductCategoryType).map((key, index) => (
+                <SelectItem key={index} value={ProductCategoryType[key]}>
+                  {key}
+                </SelectItem>
+              ))}
+            </SelectGroup>
+          </SelectContent>
+        </Select>
+      </div>
+    </div>
+  );
+}
