+++ src/components/projectWorks.tsx
@@ -4,7 +4,7 @@
  * © Copyright Utrecht University (Department of Information and Computing Sciences)
  */
 import * as React from "react";
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
+import { CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import AddProductModal from "@/components/addProductModal.tsx";
 import EditProductModal from "@/components/editProductModal.tsx";
 import { useContext } from "react";
@@ -15,26 +15,9 @@ import {
   ProjectDTO,
 } from "@team-golfslag/conflux-api-client/src/client";
 import { Button } from "@/components/ui/button";
-import {
-  Tooltip,
-  TooltipProvider,
-  TooltipTrigger,
-  TooltipContent,
-} from "@/components/ui/tooltip.tsx";
 import { ApiClientContext } from "@/lib/ApiClientContext.ts";
-import { Edit, Pencil, Trash2, X } from "lucide-react";
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-  AlertDialogTrigger,
-} from "@/components/ui/alert-dialog.tsx";
-import { Badge } from "@/components/ui/badge.tsx";
+import { Edit, X } from "lucide-react";
+import ProductCard from "@/components/productCard.tsx";
 
 type ProjectWorksProps = { project: ProjectDTO };
 
@@ -129,96 +112,17 @@ export default function ProjectWorks({ project }: Readonly<ProjectWorksProps>) {
             {project.products
               ?.filter((p) => p.categories.some((c) => c === catType))
               .map((product) => (
-                <Card className="flex flex-col gap-1 border border-gray-200 p-3 shadow-sm">
-                  <div className="flex items-center justify-between">
-                    <p className="font-semibold">{product.title}</p>
-                    {editMode && (
-                      <div>
-                        <TooltipProvider>
-                          <Tooltip>
-                            <TooltipTrigger asChild>
-                              <Button
-                                variant="ghost"
-                                size="sm"
-                                className="p-0 text-blue-500 hover:text-blue-700"
-                                onClick={() => {
-                                  setIsEditModalOpen(true);
-                                  setEditedProduct(product);
-                                }}
-                              >
-                                <Pencil className="h-4 w-4" />
-                              </Button>
-                            </TooltipTrigger>
-                            <TooltipContent>Edit Product</TooltipContent>
-                          </Tooltip>
-                        </TooltipProvider>
-                        <AlertDialog>
-                          <TooltipProvider>
-                            <Tooltip>
-                              <TooltipTrigger asChild>
-                                <AlertDialogTrigger asChild>
-                                  <Button
-                                    variant="ghost"
-                                    size="sm"
-                                    className="text-destructive hover:text-destructive/80 p-0"
-                                    onClick={() => {
-                                      setIsDeleteDialogOpen(true);
-                                      setDeletedProduct(product);
-                                    }}
-                                  >
-                                    <Trash2 className="h-4 w-4" />
-                                  </Button>
-                                </AlertDialogTrigger>
-                              </TooltipTrigger>
-                              <TooltipContent>
-                                Delete contributor
-                              </TooltipContent>
-                            </Tooltip>
-                          </TooltipProvider>
-                        </AlertDialog>
-                        <AlertDialog
-                          open={
-                            isDeleteDialogOpen &&
-                            deletedProduct?.id === product.id
-                          }
-                          onOpenChange={(isOpen) => {
-                            if (!isOpen) {
-                              setIsDeleteDialogOpen(false);
-                              setDeletedProduct(null);
-                            }
-                          }}
-                        >
-                          <AlertDialogContent>
-                            <AlertDialogHeader>
-                              <AlertDialogTitle>
-                                Are you absolutely sure?
-                              </AlertDialogTitle>
-                              <AlertDialogDescription>
-                                This will remove {product.title} from this
-                                project. This action cannot be undone.
-                              </AlertDialogDescription>
-                            </AlertDialogHeader>
-                            <AlertDialogFooter>
-                              <AlertDialogCancel>Cancel</AlertDialogCancel>
-                              <AlertDialogAction
-                                className="bg-destructive hover:text-destructive border-destructive border-1 text-white hover:bg-white/10 hover:font-bold"
-                                onClick={deleteProduct}
-                              >
-                                Delete
-                              </AlertDialogAction>
-                            </AlertDialogFooter>
-                          </AlertDialogContent>
-                        </AlertDialog>
-                      </div>
-                    )}
-                  </div>
-                  {product.url && (
-                    <p className="text-sm text-gray-600">{product.url}</p>
-                  )}
-                  <Badge variant="secondary" className="h-5 px-2 py-0 text-xs">
-                    {product.type}
-                  </Badge>
-                </Card>
+                <ProductCard
+                  product={product}
+                  editMode={editMode}
+                  setIsEditModalOpen={setIsEditModalOpen}
+                  setEditedProduct={setEditedProduct}
+                  isDeleteDialogOpen={isDeleteDialogOpen}
+                  setIsDeleteDialogOpen={setIsDeleteDialogOpen}
+                  deletedProduct={deletedProduct!}
+                  setDeletedProduct={setDeletedProduct}
+                  deleteProduct={deleteProduct}
+                />
               ))}
           </>
         ))}
